+++ ldapdao/plugins/auth/ldapdao/ldapdao.auth.php
@@ -405,6 +405,14 @@ class ldapdaoAuthDriver extends jAuthDriverBase implements jIAuthDriver
             return '%%'.$val.'%%';
         }, $searchStr);
         $values = array_values($userLdapAttributes);
+        // escape parenthesis
+        $values = array_map(function ($val) {
+            return str_replace(
+                array('(', ')'),
+                array('\\(', '\\)'),
+                $val
+            );
+         }, $values);
         $values[] = $userDn;
         $values[] = $login;
         $values[] = $login;
@@ -417,7 +425,7 @@ class ldapdaoAuthDriver extends jAuthDriverBase implements jIAuthDriver
         $grpProp = $this->_params['searchGroupProperty'];
 
         $groups = array();
-        $search = ldap_search(
+        $search = ldap_search (
             $connect,
             $this->_params['searchGroupBaseDN'],
             $filter,